pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.0",
    "flask-cors>=5.0.1",
    "flask-session>=0.8.0",
    "flask-sqlalchemy>=3.1.1",
    "pyjwt>=2.10.1",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.0",
    "sendgrid>=6.12.0",
    "sqlalchemy-serializer>=1.4.22",
    "werkzeug>=3.1.3",
]
